I am a bad person.
I was fed up of seeing these messages in the logs for the `bosh-agent` unit
tests. I looked into switching to the new style of using the YAGNATS client but
after discussing it we decided that it looks like a considerable amount of work
that we don't want to do right now.

Since the BOSH team is responsible for YAGNATS there is no risk of us deleting
the old way of doing things before this has been changed.

Like I said, I'm a bad person.